Add Zone.Identifier metadata for _Web-C1_Core_K.png sponsor image
