+++ agenda/src/assets/sponsors/01_nulogo_the-purple.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1t9tAfyPTKHUkbaNw34NSGpFb-gzZ0xVb&export=download&authuser=3&confirm=t&uuid=be95a1bd-8bd4-42c9-9066-9647a9f3ccad&at=ALoNOgkLmA7OaFKp5puzJm1ry2Dv:1749534197268
