+++ agenda/src/assets/sponsors/1-Original.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1PQGiaZ16sKQE_m9LUmrDfddlZHDmuGbS&export=download&authuser=3&confirm=t&uuid=ce41a2f8-2f23-46ea-b1e9-0000379e6c1b&at=ALoNOgmPRrPssMWTYN69VKMz7Yyc:1749533952992
