+++ agenda/src/assets/sponsors/AZ_RGB_H_COL.PNG:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1d9BYClxpVbcN4asGXt-Q_iSfFZu7dlKv&export=download&authuser=3&confirm=t&uuid=06854b91-1cc2-45ea-9d46-dee7a4552b5e&at=ALoNOgmGBVHJjXwz9R5xY_Y09ZSy:1749533633197
