+++ agenda/src/assets/sponsors/Caylent Logo Black.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1ACZmFAXHKu8m1F3p1gMFJzEpPAmEnL1u&export=download&authuser=3&confirm=t&uuid=1bb8f605-ba2e-455e-87f7-3bba0e900acf&at=ALoNOgkBA0Wd9149d9L8Oph6fODE:1749533737519
