+++ agenda/src/assets/sponsors/DoiT_Logo_Ink_Spark.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1pbW5VVJ5BF3MsgWZsdsEku_LvnvvefBW&export=download&authuser=3&confirm=t&uuid=64ed04ab-ec5e-4399-9720-8549c3ea2747&at=ALoNOglrZTl713-OePWoWCHY9g2h:1749533900593
