+++ agenda/src/assets/sponsors/EPAM_LOGO_Black.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1IZnJDbzMXcgzIti_ekDYY_OW9TLGnFby&export=download&authuser=3&confirm=t&uuid=a2beaa0a-875c-41f0-a3a9-d5e08538dfc1&at=ALoNOgnMjeGuvK-BKpn73ivm1A0D:1749533942301
