+++ agenda/src/assets/sponsors/Grafana_Logo_Stacked_FullColor_dark (1).png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1zepd98X981tKuSAubsdQWpOHf_Y7Jl7w&export=download&authuser=3&confirm=t&uuid=845122f7-e539-45e6-8f61-d216e0582cff&at=ALoNOgkRbu77YEW3-xQnritKX46Q:1749533974986
