+++ agenda/src/assets/sponsors/IBM_logo®_pos_blue60_RGB.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1zS-8JXHsdmV-pX_Taztp-9rFWhO97Omu&export=download&authuser=3&confirm=t&uuid=c5b80827-2036-4e53-838e-f7681b33cf6b&at=ALoNOgmaTrEwsRq5za0_joq28JN_:1749534153977
