+++ agenda/src/assets/sponsors/Main_Logo_digital (1).png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1WGQQe-Ui_GyHFIg31mbrCJVZSyvL98eS&export=download&authuser=3&confirm=t&uuid=2d5d4f73-ba46-45f5-afa0-011338c4169c&at=ALoNOgkj2a2-jikKp1iND1d1RP-9:1749534277250
