+++ agenda/src/assets/sponsors/SoftServe logo.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=1BFHth-cAzHYeonfbZmXV169cpn53Fcve&export=download&authuser=3&confirm=t&uuid=db0f96a3-c9c8-4b9b-9fab-bc0aa25c9350&at=ALoNOgklu74X9TNotcUMWQTp418N:1749534216302
