+++ agenda/src/assets/sponsors/Uala_Logo_Lockup_Horizontal_FondoBlanco.png:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://convertio.co/
+HostUrl=https://s175.convertio.me/p/vwO6W4jwVBKVJTVnf7NA4A/876b563d4a46c55a6c9ac162b03ac90c/Uala_Logo_Lockup_Horizontal_FondoBlanco.png
