+++ agenda/src/assets/sponsors/_Web-C1_Core_K.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://drive.usercontent.google.com/download?id=15Sgdx8VRWMwHKyxzRNjXxHxlhxiiYDBk&export=download&authuser=3&confirm=t&uuid=9e183b6c-16e1-4825-8253-b5917d18c8c8&at=ALoNOgm1Txf120BL1rysbjmiyHbN:1749533727712
